requirements.txt
pandas
alpaca-py
yfinance
matplotlib
psycopg2
fastapi 
uvicorn
strawberry-graphql
python-dotenv
sqlalchemy
pydantic
numpy
python-jose[cryptography]